[StoreBundle, ThemeBundle] introduce separate theme-bundle which uses pimcore-sites to define the themes. StoreBundle than decorates that and changes that to CoreShop stores. ThemeBundle can now be used for Multisite Projects that have to have multiple themes.
