+++ src/CoreShop/Bundle/StoreBundle/Resources/config/pimcore/config.yml
@@ -9,13 +9,9 @@ pimcore:
                     coreShopStore: CoreShop\Bundle\StoreBundle\CoreExtension\Store
                     coreShopStoreMultiselect: CoreShop\Bundle\StoreBundle\CoreExtension\StoreMultiselect
 
-
 jms_serializer:
     metadata:
         directories:
             coreshop-store:
                 namespace_prefix: "CoreShop\\Component\\Store"
                 path: "@CoreShopStoreBundle/Resources/config/serializer"
-
-liip_theme:
-    load_controllers: false
