+++ src/CoreShop/Bundle/StoreBundle/Resources/config/services/theme.yml
@@ -3,21 +3,15 @@ services:
         public: true
 
     coreshop.store.theme_resolver:
-        class: CoreShop\Bundle\StoreBundle\Theme\ThemeResolver
+        alias: coreshop.theme.resolver
+
+    coreshop.store.theme_helper:
+        alias: coreshop.theme.helper
+
+    coreshop.theme.store_resolver:
+        class: CoreShop\Bundle\StoreBundle\Theme\StoreThemeResolver
+        decorates: 'coreshop.theme.resolver'
         arguments:
             - '@liip_theme.active_theme'
             - '@coreshop.context.store'
             - '@coreshop.repository.store'
-
-    coreshop.store.theme_helper:
-        class: CoreShop\Bundle\StoreBundle\Theme\ThemeHelper
-        arguments:
-            - '@coreshop.store.theme_resolver'
-            - '@liip_theme.active_theme'
-
-    coreshop.listener.theme_request:
-        class: CoreShop\Bundle\StoreBundle\EventListener\ThemeRequestListener
-        arguments:
-            - '@coreshop.store.theme_resolver'
-        tags:
-            - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest }
