+++ src/CoreShop/Bundle/ThemeBundle/Resources/config/services.yml
@@ -1,0 +1,20 @@
+services:
+    _defaults:
+        public: true
+
+    coreshop.theme.resolver:
+        class: CoreShop\Bundle\ThemeBundle\Service\ThemeResolver
+        arguments:
+            - '@liip_theme.active_theme'
+
+    coreshop.theme.helper:
+        class: CoreShop\Bundle\ThemeBundle\Service\ThemeHelper
+        arguments:
+            - '@coreshop.theme.resolver'
+            - '@liip_theme.active_theme'
+
+    coreshop.theme.listener:
+        arguments:
+            - '@coreshop.theme.resolver'
+        tags:
+            - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest }
