+++ src/CoreShop/Bundle/ThemeBundle/Resources/pimcore/config.yml
@@ -1,0 +1,2 @@
+liip_theme:
+    load_controllers: false
