+++ src/CoreShop/Bundle/ThemeBundle/composer.json
@@ -1,7 +1,7 @@
 {
-  "name": "coreshop/store-bundle",
+  "name": "coreshop/theme-bundle",
   "type": "pimcore-bundle",
-  "description": "CoreShop - Store Bundle",
+  "description": "CoreShop - Theme Bundle",
   "keywords": [
     "coreshop",
     "pimcore",
@@ -18,10 +18,7 @@
     }
   ],
   "require": {
-    "coreshop/store": "^2.0",
-    "coreshop/resource-bundle": "^2.0",
-    "coreshop/currency-bundle": "^2.0",
-    "coreshop/theme-bundle": "^2,0",
+    "liip/theme-bundle": "^1.4",
     "pimcore/pimcore": "^5.4.0"
   },
   "config": {
@@ -29,7 +26,7 @@
   },
   "autoload": {
     "psr-4": {
-      "CoreShop\\Bundle\\StoreBundle\\": ""
+      "CoreShop\\Bundle\\ThemeBundle\\": ""
     }
   },
   "minimum-stability": "dev",
